src/main/java/org/quickstart/servlets/servlet/HomeServlet.java
package org.quickstart.servlets.servlet;

import org.quickstart.config.JSPConfig;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

// Main page - see web.xml
public class HomeServlet extends HttpServlet {
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        request.getRequestDispatcher(JSPConfig.PATH + "index.jsp").forward(request, response);
    }
}
